store is updated
